tray: wrap a bunch of errors
I know that I should have done this originally, but I had pretty much
no idea how the protocol worked when I started so the code changed so
many times that wrapping a return from a function that was experimental
and quite probably incorrect was not particularly feasible.

There are still more things that need to be wrapped, but this should be
most of the big ones.